An attempt at game of life in powershell
This is my fist attempt at simulating the game of life in powershell.
There are certain problems with the code currently. One of them is
accepting console input from mouse. Secondly, the code for generating
the next generation is taking about 10.5 secs..which is too long. Need
to work on that well.